requirements.txt
django==4.2.*
fastapi
channels[daphne]
channels_redis
dj-database-url
python-dotenv
whitenoise
ASGIMiddlewareStaticFile